[MOVE/#206] 마이페이지 프로필 파일 위치 이동
