+++ core/src/main/java/com/terning/core/designsystem/component/item/ProfileWithPlusButton.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.onboarding.signup.component
+package com.terning.core.designsystem.component.item
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
@@ -12,14 +12,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.terning.core.R
 import com.terning.core.designsystem.component.image.TerningImage
-import com.terning.feature.R
 
 @Composable
-fun SignUpProfile(
+fun ProfileWithPlusButton(
     index: Int,
     modifier: Modifier = Modifier,
 ) {
@@ -37,7 +36,7 @@ fun SignUpProfile(
     ) {
         Image(
             painterResource(id = grade),
-            contentDescription = stringResource(id = R.string.sign_up_profile_image),
+            contentDescription = null,
             modifier = modifier
                 .clip(shape = RoundedCornerShape(76.dp))
                 .size(80.dp)
@@ -54,6 +53,6 @@ fun SignUpProfile(
 
 @Preview(showBackground = true)
 @Composable
-fun SignUpProfilePreview() {
-    SignUpProfile(index = 1)
+fun ProfileWithPlusButtonPreview() {
+    ProfileWithPlusButton(index = 1)
 }
