+++ core/src/main/res/drawable/ic_sign_up_button.xml
